+++ packages/core/src/http/cookie.ts
@@ -1,5 +1,5 @@
-import type { Logger } from '@faasjs/logger'
 import { deepMerge } from '@faasjs/node-utils'
+import type { Logger } from '@faasjs/node-utils'
 import { Session, type SessionOptions } from './session'
 
 export type CookieOptions = {