+++ packages/core/src/http/session.ts
@@ -1,5 +1,5 @@
 import { createCipheriv, createDecipheriv, createHmac, pbkdf2Sync, randomBytes } from 'node:crypto'
-import type { Logger } from '@faasjs/logger'
+import type { Logger } from '@faasjs/node-utils'
 import type { Cookie } from './cookie'
 
 export type SessionOptions = {